Bugfix Exeption class and Debugging support
This patch bugfixes a pointer management within the Exeption class,
where a shared_ptr is required instead of the unique_ptr.

Moreover, the debugging support has been imporved with the introduction
of a new DEBUG macro (defined into object.h). Using this macro allows to
strip debugging sentences when a "Release" version is built.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>