Bugfixed static plugin auto-registration
This patch allows a static plugin to auto-register itself tanks to the
proper exploitation of the StaticPlugin class constructor.

Since the static plugin is build within a static library (.a archive),
at link time, only reference symbols are linked by default.
Thus, the statically allocated object is normally discarded.
The "whole-archive" linker flags forces the linking of all symbols which
allows the self-registration mechanisms to work still using a static
library (which is better for the CMake code organization).

Interesting references:
http://stackoverflow.com/questions/1804606/static-initialization-and-destruction-of-a-static-librarys-globals-not-happening
http://www.mail-archive.com/cmake@cmake.org/msg01890.html

NOTE: The same "self registration" style is also possible in plain-C,
e.g. with the __attribute__((constructor)) GNU extension.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>