Object model refactoring
This patch introduce a new (single) source file for the definition of
all barbeque modules interface. The bbque/modules.h header will provide
the well defined interface for each module that could be instantiated
within the barbeque tool.

The TestModuleIF abstract class defines a single and simple interface to
develop testing modules for other barbeques subsystems. This interface
requires the definition of a Test method. More than one plugin (either
static or dynamic) could be defined to support unit testing.

A dummy module implementation is also provided, implementing the
TestModuleIF. This simple module is provided just as a reference design
for the development of other modules.

TODO: more modules interfaces will be defined and added to the modules
header file.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>