Added ResourceManager and PlatformServides
This patch provides a _stub_ implementation of two main barbeque modules.

The ResourceManager class will provides the brain to barbeque. It will
host all the glue code among other modueles and define the running
policy. It provides the Go method which is the actual barbeque entry
point.

The PlatformServices class provides an implementation for the service
invoke methods exposed to modules. This class will host the definition
of supported services and the corresponding struct based parameters.
NOTE: the definition of supported services should be C/C++ compatible.

The main method will be in charge of mere initializations and
housekeeping activities, such as parsing the command line, intializing
core modules, start the games by calling the Go method of the
ResourceManager and finally cleanying up everything once a termination
request occours.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>