Added namespace based modules name resolution
This patch allows to request the plugin manager for a module using a
namespace based naming. Each module should be named according to a
proper namespace, e.g. test.<module_name> for each module
<module_name> which implements the Test interface.

Once a new object is required to the plugin manager, it will return the
nearest matching module belonging to that class. Thus, for instance a
request for a "test." module could match either "test.console" or
"test.log4cpp".

If no plugins are registered within this namesapce, the object will not
be created, and a null pointer returned. If an object with the exact
name could be found, that object will be returned.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>