Added support for plugins object init data
This patch extends the PF_ObjectParams to host a generic void pointer to
a plugin defined structure. Each module interface should define the
struct that can be associate to this attribute.

The new extension is used to pass the Category string to the LoggerIF
modules. A pluging implementing this interface, such as the Log4CPP
logger, could use this information to properly configure its behaviors.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>