Moved plugin configuration at object creation time
This patch move the plugin configuration setup at object creation time.
The first time an object of this class is built, we load the
configuration from file.

This approach allows to exploit the PlatformSerivces class. Indeed, at
plugin registration time (in the case of a static plugin) the
InvokeService has not yet been properly configured. Insteas, when an
object of this plugin is built for the first time, everyting is on place
to correctly interact with the configuration manager and collect
configuration options required by the plugin.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>