Added a platform service to load configuration
This patch provides the support for modules to load their configuration
options from a file. The PlatformService class defines a new
PF_SERVICE_CONF_DATA service which exploit the Boost "program options"
interface for the definition and collection of configuration options
from a file and/or command line.

A module could define its specific options, along with default value and
containers, and then invoke the platform service to ask for a
configuration update. The Log4Cpp logger has been already integrate with
this new interface to asses both a working proof and to collect the
location of its configuration file.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>