Addded dynamic plugin examples
This patch provides a better (and working) example of static/dynamic
plugin code. The DummyTest Object is used to demo how to code a plugin
which could be compilerd either statically or dynamically.

Macros are used to define at compile time two different class scopes for
static and dynamic objects. A new CMake targe allows to compile the same
source both as static plugin (with auto registration support) and
dynamically, to be loaded from the plugin directory.

A better CMake support should be provided to define which modules to
compile statically and which should produce a .so module.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>