Update plugin manager to support pure CPP modules
This patch slightly change the plugins manager interface to provide a
better support for CPP only coded plugins. Indeed, these plugins do not
require an object adapter, thus we admit the possibility to specify a
NULL pointer adapter (instead of a reference). In this case, if the
plugin is C coded and a NULL object adapter is passed, the object
creation will be avorted, viceversa, if the module is CPP coded,
everithing goes on as before.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>