Added template implementation for YaMCA module
This patch add the template code to develop the YaMCA resource manager
heuristic. The provided code is just enough to build the plugin to
produce a shared object (.so).

All the required heuristic code should be added and the entry point is
just the "schedule" method. Console output are also added just for
template testing and must be removed before releasing the code.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>