Added first RTLib stub code
This patch provides an initial definition of the run-time library
itnerface and a stub code for the library implementing it.

The designed interface is quite similar to that used by plugins with
these main differences:
* plugins libraries are now representd by applications
* plugins objects are now represented by "execution contexts"

This schema allows for single applications to register more than one
"execution context", this can be the case of complex applications which
are based on a collection of stream kernels to be concurrently scheduled
on the computation fabric.

TODO:
- review the interface with application developers, during the next
  2PARMA telco;
- provide the implementation of at leat one communication channel

Signed-off-by: Patrick Bellasi <derkling@gmail.com>