Plugin Data: parsing of plugin specific data
Added the support for parsing plugin specific data from the recipe.

A plugin could need some specific information in order to run correctly.
For instance an Optimizer module, in its scheduling process, may save
specific information upon a given Application or WorkingMode that it aim to
retrieve next time that application has to be scheduled again.

The idea is to store each data in a map of key-value pairs of type
string. Such approach has chosed in order to have as flexibility as
possibile.

Let's think if the Optimizer module want to keep track of the number of
times an application as been re-scheduled (for a purpose that it only
cares). The Optimizer can add its plugin info to the Application object
and then set a pair <"num_of_sched", "7">.

Classes PluginData, PluginDataContainerIF, PluginDataContainer define
and implement the support.

Signed-off-by: Giuseppe Massari <joe.massanga@gmail.com>
Signed-off-by: Patrick Bellasi <derkling@gmail.com>