ResourceAccounter and WorkingMode
The ResourceAccounter component has been added. It provides methods for
reading or updating resources runtime information. Basically information
managed are: availabilities, amount of total and used resources,
applications using the resources.

Each resource in the system has its own descriptor, which is stored in a
tree based structure, implemented by ResourceTree. Accessing resource
descriptors it's a matter of pattern matching in a namespace-like style.
This means that we refer to an application as the following example:
"arch.clusters.cluster2.pe1.mem0".
Two kinds of search are possibile:
	1) Exact match: we want a well defined system resource.
	2) Template path match: we are just interested in knowing if
	there are any resources compatible with the path specified. For
	instance, considering the example above, the template path is:
	"arch.clusters.cluster.pe.mem".

A WorkingMode is basically a set of resource usages, specified in the
recipe used by the application in order to get the expected QoS. The QoS
is express by the WM "value".
In the working mode descriptor we keep track of the transition overheads
too. Such overheads occour when an application switches between
different working modes
(TransitionOverheads).

Signed-off-by: Giuseppe Massari <joe.massanga@gmail.com>
Signed-off-by: Patrick Bellasi <derkling@gmail.com>