Added first RPC stub code for a FIFO based channel
This patch provides an initial definition of a FIFO based message
passing communication channel.  Baed on this implementation,
applications and the Barbeque RTRM communicate over a pair of FIFO
channels, by exchanging a set of well defined messages.

The set of messages exchanged has been defined with the rpc_messages.h
header file. This set of messages is well suited also for the
implementation of a SOCK based communication channel.

Both Barbeque and the RTLib should implement this communication
protocol, to this purpose both a server and a client interface have been
defined by rpc_fifo_server.h and rpc_fifo_client.h.  The former (i.e.,
the server interface) is expected to be implemented as a Barbeuqe
module, while the latter (i.e., the client interface) should be
implemented as a source module to be "conditionally" linker with the
RTLib.

Right now we opted for conditional compilation of the client interface
in order to produce a small foot-print library which provides just the
support required by the target platform. Indeed, we expect to not have
more than one communication channel on the same target system.  Whenever
this assumption should be a limiting, the current design allows for a
more run-time modular approach by refining the factory method provided
by the base class.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>