Plugin Manager: dlsym clean compilation
This patch make use of a data structure to export in a "compiler
friendly" way the plugins entry point. Indeed, due to a weakness of the
ISO C standard, there is no valid cast between pointer-to-data and
pointer-to-function, thus a dlsym cast to a function pointer will
result (at least) on a compiler warning.

Indeed, the ISO C standard does not require that an object of type void*
can hold a pointer to a function. Thus, compilers conforming to the ISO
C standard are required to generate a warning if a conversion from a
void* pointer to a function pointer is attempted (for more details,
see the rationale section of the dlsym manpages).

Since we are designing the interface of the dynamically loaded object
it's possible to avoid the problem entirely by only exposing the
supported kind of pointers. Thus the solution has been that to define a
struct of function pointers, which actually defines only a member to
point the plugins entry point. The plugin manager could use the dlsym to
get a valid pointer-to-data to such a struct and then it accesses its
member to get the actual function pointer.

A convenient macro PLUGIN_INIT has also been provided which can be used
(only) by dynamic plugins to define and initialize the data structure.
See dummy_plugin.cc for an example.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>