Removing _some_ GCC flags for sections handling
This patch remove the usage of section based compilation for the
building of the Barbque core. Indeed, this flags should be used only
when a monolitic build profile will be defined otherwise we risk to
remove core unused symbols which are still required to successfully load
(and link) some modules at run-time.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>