Bugfix: segfault on logger module load errors
The logger module is a critical component of the Barbque framework.
If a logger module fails to load, or to configure, we still need to go
on without returning a null logger pointer to requesting modules.

This patch provides a simple and dummy console based logger. In case of
load/configuration problems of the logger module, the module factory
fall-back to the usage of this dummy console based logger.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>