Application: added LowValueAWM(), HighValueAWM()
A pair of methods have been added to ApplicationStatusIF (and thus to
Application) in order to get the descriptors of the working modes with
the lowest and the highest value.
The list of working modes enabled (enabled_awms) is reordered by WMs
value every time it is modified (new application starting, assertion of
constraints). Thus the methods added are simple inlined calls to front()
and back() of the std::list.
The sort of "enabled_awms" uses the function CompareWMsByValue() just
added.

The list now is initialized in SetRecipe(), which makes more sense than
doing a check on its (and constraints list) emptyness every time
WorkingModes() is called.

The data type used for the value of the working mode is 'uint16_t'
instead of the previous 'uint8_t'. This avoid ambiguity in debug
messages, where implementations of uint8_t as char could determine the
printing of unclear information.

CoreInteractionsTest plugin: added code to show up the lowest value
working mode, plus some timing messages.

Signed-off-by: Giuseppe Massari <joe.massanga@gmail.com>