RTLib API - First review
This patch add the modifications the the RTLib API identified after the
first internal review. These are the main updates:

- Synchronization type: applications could assert a synchronization type
to be either stateless or statefull, this info could be used to better
support the reconfiguration/migration strategy

- Constraints: we provide the possibility to enable/disable each single
working mode, as well as the possibility to defined boundary based
constraints. Moreover, a new parameter of the RTLIB_Constraint data
structure allows to specified if the corresponding constraint should be
added or removed

- Working Mode: the SetWorkingMode call has been replaced by a
GetWorkingMode, which is now a blocking call exposed to the
application. An application could easily block on this call waiting for
an AWM being assigned by the RTRM.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>