ResourceAccounter: query state by path template
A new feature has been added to the Resource Accounter. It's now
possibile to query the state (available, used and total amounts) not
only of a specific resource, but of a set of resources too. The set is
referenced by its path template.
Thus for instance, if we want to know how many processing elements are
there on the platform installed we should pass to method Total() a path
template as "arch.clusters.cluster.pe".

For this feature a new utility function has been insert into utility.h:
IsPathTemplate(). Such function get a resource path as input and returns
a boolean saying if the path is a template or a specific resource path.

ResourceAccounter.RegisterResource() now returns an exit code
(RegExitCode_t) instead of a void. This makes the method more readable
and more informative.

utility.h: Function ConvertValue() has been rewritten in a more
efficient and readable form.

Comments have been added into ResourceAccounterConfIF in order to give a
richer explaination about methods.

Signed-off-by: Giuseppe Massari <joe.massanga@gmail.com>