ResourceAccounter: Added clustering factor
The clustering factor is a potential useful information for the
scheduling/optimizing algorithm. Basically it is the number of cluster
of the system architecture.
Such information is returned by ClusteringFactor() in ResourceAccounter,
or ResourceClusterFactor() in SystemView.
Given the resource path as input these calls return 1 if the resource is
system-wide, 0 if the resource path doesn't match any resource, and the
clustering factor computed (once) if the resource is a clustered one
(i.e a processing element PE).

Signed-off-by: Giuseppe Massari <joe.massanga@gmail.com>