FIX: using Boost Filesystem V3
This small patch force the usage of version 3 for the Boost Filesystem
library. Indeed, version 2 is deprecated, and will not be included in
Boost releases 1.48 and later.

To ease the transition, Boost releases 1.44 through 1.47 will supply
both V2 and V3. However, version 2 is the default version for Boost
release 1.44 and 1.45. Version 3 will be the default starting with
release 1.46.

More details at:
http://www.boost.org/doc/libs/1_46_1/libs/filesystem/v3/doc/index.htm

Signed-off-by: Patrick Bellasi <derkling@gmail.com>