UPDATE: Removing logger::Debug on release compile
This patch make use of the BBQUE_DEBUG macro to ensure the removal of
logger::Debug calls once compiling release versions.
Indeed, in release compilation the Debug method is not virtual and has
an empty implementation, thus the compiler support us by removing the
unnedded call from the generated code.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>