ResourceTree: added "hybrid path" search
As "hybrid path" we refer to a resource path which is part template and
part ID-based.
This form of resource path is needed when for instance we need to make
query as "How many Processing Elements are there in the Cluster 1?".
Thus such query implies a unique call to
ResourceAccounter.Availability(path), where the path should be:
"arch.clusters.cluster1.pe".
This can be splitted in "arch.clusters.clusters1" (ID-based part), and
"pe" the template part.
The alternative would be to make "N" query, one for each of the
processing elements contained in cluster 1 (pe0, pe1, ...) using their
ID-based path.

In order to allow this kind of query from ResourceAccounter, the method
findSet() has been added to ResourceTree.

TODO: [ResourceAccounter] Review queryState() to manage more effectively
the case of specific resource query.

Signed-off-by: Giuseppe Massari <joe.massanga@gmail.com>