Resource: added Acquire() and Release()
Methods AddUsed() and SubUsed() have been removed from ResourceState
class, and replaced by Acquire() and Release() in class Resource.

Calling the first, an application acquires a required amount of
resource (if available). Thus the Resource object keeps track of the
application (PID) and the amount of resource it holds.
A call to Release() frees the resources held by the application.

ResourceUsage:
Attribute "bind_path" has been replaced by a list ("binds"), which is
filled when the resource binding is solved (by the Scheduler /
Optimizator), with all the descriptors of the resources matching the
bind defined.

WorkingMode:
ResourceUsage() renamed as ResourceUsageValue() in order to avoid
ambiguity in code reading.

Signed-off-by: Giuseppe Massari <joe.massanga@gmail.com>