ResourceAcc: new resource reservation mechanism
The resource reservation mechanism has been completely changed.
What happen now is that calling SetNextSchedule() on Application
triggers immediately the request of resources. This is provided by the
new method AcquireUsageSet() of ResourceAccounter.
Here, the R.A. checks if the Application making the acquire request,
holds yet another resource set. If so, it first release the current set
(ReleaseUsageSet()) and then start with the reservation of the resources
specified in the new one.
The new resource set is the resource usage set defined in the working
mode scheduled (the one returned by NextAWM() of Application).

Before succesfully completing the resource acquisition/reservation
procedure, the R.A. checks a couple of things on all the resource usage
requests:
1) A resource binding has been defined ?
2) The request is satisfiable? (enough resources available)
If a check fail, the acquire request (and thus the working mode
scheduled) is rejected.

Accordingly the Application doesn't commit the scheduled working/mode
status. Conversely, if the request is accepted the Application flags
itself as "marked to switch". Such flag tells to the Scheduler/Optimizer
that its choice upon the application is currently schedulable.

Signed-off-by: Giuseppe Massari <joe.massanga@gmail.com>