WorkingMode: Added BindResources()
Resource paths taken from the recipes use an ID number agnostic with
regards to the one actually used by Barbeque for the real system
resources registered. Thus a binding must be solved (reasonably by the
Scheduler/Optimizer module) in order to make the request of resources
satisfiable. For doing that, a new method has been added to class
WorkingMode: BindResources().

The method takes the resource name we want to bind (i.e. "cluster"),
the ID of the destination system resource, and the source resource ID
(optional). Then it builds the "bind path", by substituting
"cluster+[source ID]" with "cluster+[dest ID]". Once done,
GetResources() from ResourceAccounter is called to retrieve the list of
descriptors matching the bind path.

If the source ID is left blank (or similarly equal to -1), the method
will bind ALL the resource path containing "cluster" or "clusterN" to
the descriptor referenced by such path, with the descriptor resource ID
of the system.

Signed-off-by: Giuseppe Massari <joe.massanga@gmail.com>