Added first implementation of the ApplicationProxy
This patch provides the first implementation of the ApplicationPoxy.
This is the Barbque component in charge to provide the unified interface
to access applications. This component is in charge to setup and manage
communication channels: both the Bbque input channel and these specific
to each registered application.

In order to improve efficiency by better exploiting host side
parallelism, this compoent takes care to build (concurrent) execution
context for each communication needs between the RTRM and the
applications. Indeed, a specific thread could be generated to manage
separately each incoming/outgoing request. This design allows thereby to
better exploit (RT) threads attributes for instance to boost
high-priority messages by setting a different thread priority.

The implementaiton is just embrional, providing support up to know just
for few connection setup/teardown messages and only for application
originated messages. Specifically, the implementation for Bbque
originated messages is still missing.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>