Bugfix to handle channel build failure
When the RPC module is not able to build a communication channel (e.g.
too many files open), the Application Proxy silently ignore the
application request.

The application could/should try again the PAIRING after a timeout has
expired. Right now we don't provide this feature and the application
simply fails the RTLib initialization if the BBque server does not
repsond within a 250ms timeout.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>