[AP] Refactoring Execution Ctx to Session Ctx
This patch refactors the Execution Context by making it a base class
from with to derive the Request Context class (rqsCtx_t). This base
class will be used also to derive a Transaction Context class used to
manage Bqrbeque requests to applications.

This design allow also to keep a single map of active Execution Contexts
which give an overall view on the on-goning communications.

Moreover, to avoid a concept overlapping with the application specified
Execution Contexts, now we use the therm "Session Context" (snCtx_t) to
represent a thread handling a communication with the application. These
threads could have two specializations:
- "Request Sessions" (rqsSn_t) are snCtx_t initiated by an application
  sending a message to the RTRM
- "Command Sessione" (cmdSn_t) instead are snCtx_t initiated by Barbque
  contacting an application.

The support for the second kind of sessions (comSn_t) will be provided
by a following patch.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>