[ApplicationManager] Refactoring StartApplication
This patch refactors the StartApplication method to improve code
readibility, by isolating the recipe load code into a new private
LoadRecipe method.

TODO: the Recipe class should be refactored to becoms a pure Application
client class. Indeed, the current design pass an Application pointer to
the LoadRecipe method which is used to populate this application with
recipe defined static constraints.

This is a poor design, maybe also inefficient when multiple applications
share the same recipe. It would be better to refactor this way:
- parse static contraints and save them locally to the recipe object;
- provida a new Recipe::GetStaticContraints method which return the list
  of static contraint
- uss the Application initialization code to both load the recipe and
  get the list of static constratins to do whatever is required by the
  Application code itself.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>