[ApplicationProxy] WIP: Added Command Session
This patch provide an initial support for Command Sessions.
Right now we implement just the StopExecution session for a specified
(Application,ExecutionContext) pair.

TODO: add support to the RTLib for this command

TODO: the timeout value sould be better defined (actually it should be
correctly implemented). We should used std::chrono in the Barbque side
but translate this to a "struct timeval" at RTLib side.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>