[Application] Added AppPid_t type
This patch make use of an AppPid_t type to represent the PID of an
application. This type should be an alias to pid_t, but since this seem to
give some problems with boot::program_option when parsing the value from
the configuration file, right now we simply use a generic uint32_t base type.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>