[Application] Improved tracing and debugging
This patch add a new method StrId() which return a formatted string
identifier for the application with this format:
<PID>:<NAME>:<EXC>
where all fields are fixed space and name is the process name truncated
at 6 chars.

Some more debugging has been added, and a minor code refactoring
for the class destructor allows to properly release resoruces.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>