[AM] Refactoring to solve endless-loop on ChangeSchedule
This patch provides a big code refactoring and cleanup, especially to
"solve" the endless-loop and data structure corruption on
StopApplication and ChangeSchedule.

The main modificaiton involves the status_vec vector, right now we
release all the resources once an application is stopped.
Some more debugging sentences have been added, exploiting the new
Application::StrId() for improved consistency on logging messages.

TODO: investigate better on Application destructor invocation, this
should be called when an application is stopped...

TODO: refactor the RecipeLoader to target ticket #2

TODO: add ExitCode for the ApplicationManager::StopApplication to target
the issues ticket #10

Signed-off-by: Patrick Bellasi <derkling@gmail.com>