[A/AM] Added DISABLED state
This patch add a new Application state which is used by newly created
EXC and EXC which are disabled by the controlling application.
The DISABLED state is now the default for newly created EXC, moreover
an application could be moved into this state when it is not required to
associate resources to it. The disabled state could be used for instance
to temporarely disable EXC.

A new Enable method has been added to move an application from DISABLED
to READY state. Vice verse, a Disable method releases all the
EXC assigned resources and move it to the DISABLE state whatever is its
original state.

TODO:
- align the Start/Stop and Enable/Disable concepts to these defined by
  the RTLIB, as by issue ticket #6

Signed-off-by: Patrick Bellasi <derkling@gmail.com>