[PluginsData] New interface for plugin specific data
For managing plugin specific data, a new class has been implemented
(PluginsData), replacing the two used before (PluginData,
PluginsDataContainer).

Basically, PluginsData shows a couple of simple methods for getting and
setting data: GetAttribute() and SetAttribute().
Data are thus stored in a multimap having the plugin name as key, and a
pair object as value. Such pair object is the attribute to be
read/written, wherein the attribute name is the key, and the value is
provided through a void pointer.

While this new approach has required much less code to be implemented,
we didn't lose flexibility. Indeed the storage of different types of
data is still allowed.

Signed-off-by: Giuseppe Massari <joe.massanga@gmail.com>