Removed Object dependency
Right now the Object based class is used to define a Logger object,
but this support has not been completely implemented.
A good implementation should provide a set of base classes which are used
to define a hierarchy to be reflected into the logging namespace.

This patch remove the Object based class dependency while a proper
integration will be developed, as defined by the issue ticked #15

Signed-off-by: Patrick Bellasi <derkling@gmail.com>