[RA] Application pointers switched to shared_ptr
A shared pointer to Application (AppPtr_t) is used in place of common
(const) pointers, for referencing set of resources usages, or the
application using a specific resource.

Moreover the Application usages maps are retrieved using the pointer as
key, instead of the Pid. This accordingly to the introduction of the
concept of Execution Context, which implies more Application descriptors
with the same Pid.

Signed-off-by: Giuseppe Massari <joe.massanga@gmail.com>