[SignalsManager] Added initial implementation
This patch provides an initial implementation for the SignalsManager class
which is in charge to define the proper support to catch system generated
signals which should trigger actions on the Barbeque framework.

Up to know we provide handling for INT and QUIT signals, which trigger
respectively a normal or abortive Barbeque termination.
The normal termination (SIGINT signal) takes care to stop all the active
applications and to clean-up all open descriptor.
Instead, the abortive termination (SIGQUIT) will produce a program
termination without releasing resources and eventually producing a dump
which could support application debugging.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>