[Recipe] Removed dependencies from Application ptr
A previous implementation of the recipe loader used a reference to the
Application descriptor for static constraint assertions and for WorkingMode
constructions. This behavior has been removed and Recipe has became a full
"client" class respect to Application.

Indeed the static constraints and the initialization of application working
modes are now managed by Application, in the method SetRecipe(), once a valid
recipe object has been instanced.

Signed-off-by: Giuseppe Massari <joe.massanga@gmail.com>