[ApplicationManager] Added schedule_vec support
This patch add a vector of application maps used to track applications
which scheduled should be changes.
Such vector could be used by ApplicationManager client modules to get a
direct reference of applications that should be reconfigured into a
specific state (e.g. applications to be stopped).

This patch prepare to ground for a new method that should be added to
this class to mark applications requiring a schedule change.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>