[ApplicationManager] Add SetSchedule
This adds a method to mark an application which needs to be reconfigured.
Once the current status of an application is changes, e.g. by a run
of the optimization policy, than it needs to be considered by the
synchronization module. To this purpose, the new method allows to mark
the application by inserting it into a proper data structure which aggregate
applications based on their new required status (e.g. to be stopped, to be
migrated, etc).

Signed-off-by: Patrick Bellasi <derkling@gmail.com>