[A/AM] Make use of SetSchedule
This patch make use of ApplicationManager::SetSchedule to move an application
from a "transitional" state to a "stable" state.

Applications are moved into a transitional state, for example, by the
optimization policy by means of a SetNextSchedule. In this case, if the
application has changed its AWM, it is also marked for a synchronization
using the SetSchedule module.

Once applications have been reconfigured, the ApplicationManager::ChangedSchedule
method is used to commit the reconfiguration request.
This method has been updated to:
- immediately updated statistics;
- this update also consider the current application state, and specifically,
  move the applications from "transitional" to "stable" states;
- remove the application from the vector of applications waiting for a
  reconfiguration, i.e. schedule_vec;

It is worth to notice that:
- ChangedSchedule works only on "stabel" states
- SetSchedule works only on "transitional" states

TODO:
- add support for EXC migration (ticket #22)

Signed-off-by: Patrick Bellasi <derkling@gmail.com>