[ApplicationProxy] Use std::map to track pending command sessions
This patch make use of a simple std::map to keep track of an initiated
command session which should generate a respose.
Enqueuing is provided by the EnqueueHandler whic _must_ be called from
the command execution context since it will take care to get the PID
of the caller to map the command session.

NOTE: according to this patch, each Barbeque originated message, i.e. each
COMMAND, _must_ use the PID of the command session thread as message token.
Indeed, the message token of the received response will be used to dispatch
it to the waiting command session.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>