[RPCMessages] Added message token
This patch add a "message token" to the RPC message header.
This token is used to match messages requests/commands with the
corresponding generated responses. Thus, an entity sending a message will
expect a response with the same token.

Right now this token is set to the message originator thread, both for
requests and command. Thus, for requests, i.e. messages originated by the
RTLIB, this is the PID of the communicaiton thread spawned by the library.

From the Barbeque side we exploit the fact that each sent command message
is generated either by a dedicated thread (e.g. for asynchronous commands)
or by the thread of a core module. Thus, for the Barbeque end-point also,
command messages tokes are set to the PID of the originating thread.

NOTE: the originating thread PID is used also to map command sessions so
that starting from the message token of a received response we are able to
dispatch it to the waiting thread which has originated the command.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>