[Application] Split State and SyncState vector
The previously used state enumeration has been updated to split the two
concepts of "current state" and "synchronization state".

Applications have a current state, which represent their current condition.
Once an event trigger an action which require to change the state of an
application, a synchronization could be required. Synchronization is required
each time the application change its current usage of resources.
For example, an applications starting or an application already running and
but willing to change the actually used resource, or again an application
running which is going to finish and releasing its resources.

A specific set of states are provided to define the possible synchronization
actions required by an application.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>