[Application] Tracking pre-sync state
This patch add support to keep track to the application state previously
to enter in SYNC state. This could be used in case of a synchronization
abort or to distinguish the actual status of an application before the
synchronization has been committed.

The SYNC_NONE constant has been defined to represent the condition
"synchronization not required".

Some minor renames and code movements are also present.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>