[ResourceAccounter] New resource booking API
Resource booking is a matter of calling BookResources(). This method,
which replace the old AcquireUsageSet(), checks the availabilities of
the resources and if the set required can be allocated, update the
accounting.

ReleaseUsageSet() has been renamed to ReleaseResources() for a more
uniform naming convention.

Moreover methods Available(), Total(), Used() have been overloaded to
allow direct query on ResourceUsage object (pointer).
This versions iterate over the Resource descriptors in the "binds" list
of ResourceUsage, accumulate, and then return, the value of the queried
attribute.

Signed-off-by: Giuseppe Massari <joe.massanga@gmail.com>