[WorkingMode] New resource binding API
The set of resources defined in the WorkingMode must be bound to a set
of system resources. This can be done by calling BindResource() and
specifying the resource name, the source ID (the one used in the
recipe), the destination ID (the one of the system resource to which
bind the request), and a map of Resource Usage objects (shared
pointers). Last argument will be filled with the ResourceUsage objects
containing the bindings (see attribute "binds").

Once the map has returned, caller can checks the availability of each
resource and thus try to set the map as the current binding for the
resource set of the working mode, by calling SetResourceBinding().

Method GetResourceBinding returns the current resource binding map.

utility.h:
Some methods have been moved and refactored as utility functions
(see AppendID(), ReplaceResourceID(), GetResourceID()).

Signed-off-by: Giuseppe Massari <joe.massanga@gmail.com>