[AM] Rename StartApplication into CreateEXC
This is the first of a series of patches intended to cleanup some code
by aligning RTLIB concepts with the Barbeque internals.

This patch provides the renaming of StartApplication with CreateEXC.
Indeed, internally to Barbeque, each single and "atomic" computational unit
which could be associated to resources is named "Execution Context" (EXC).
Each EXC could be associated to a specific recipe and each Application could
be composed or use multiple different EXC.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>